Add item event listener
